lexer refactor
