+++ src/lexer/vocabularyGenerator.ts
@@ -1,0 +1,13 @@
+import { TokenType, TokenTypeDictionary } from "chevrotain";
+
+/**
+ * Convenience function to generate a vocabulary object from a list of token definitions.
+ * @param tokens The token list to convert
+ * @returns The corresponding vocabulary
+ */
+export function genVocabulary(tokens: TokenType[]) {
+  return tokens.reduce((map: TokenTypeDictionary, obj: TokenType) => {
+    map[obj.name] = obj;
+    return map;
+  }, {});
+}
