+++ test/parser/baseParser.test.ts
@@ -1,9 +1,9 @@
 import { EarlyExitException, Lexer, NotAllInputParsedException } from "chevrotain";
-import { baseTokens, baseVocabulary } from "../../src/lexer";
+import { cronTokens, cronVocabulary } from "../../src/lexer";
 import { BaseParser } from "../../src/parser";
 
-const parser = new BaseParser(baseVocabulary);
-const lexer = new Lexer(baseTokens);
+const parser = new BaseParser(cronVocabulary);
+const lexer = new Lexer(cronTokens);
 
 function parse(input: string) {
   const lexingResult = lexer.tokenize(input);
